updated api

+++ apps/api/src/services/space.ts
@@ -2,30 +2,93 @@ import prisma from "../prismaClient";
 import { spaceMeta, speceElementMeta } from "@repo/types/src/space";
 
 export class spaceServices {
-  async createSpace(spaceMeta: spaceMeta) {
+  async createSpace(spaceMeta: spaceMeta, createrId: string) {
     try {
-      const space = await prisma.space.create({
-        data: {
-          name: spaceMeta.name,
-          backgroundImagePath: spaceMeta.backgroundImagePath,
-          width: spaceMeta.width,
-          height: spaceMeta.height,
+      if (!spaceMeta.mapId) {
+        const space = await prisma.space.create({
+          data: {
+            name: spaceMeta.name,
+
+            width: parseInt(spaceMeta.dimensions.split("x")[0] || ""),
+            height: parseInt(spaceMeta.dimensions.split("x")[1] || ""),
+            creatorId: createrId,
+          },
+        });
+
+        return space.id;
+      }
+
+      const map = await prisma.map.findFirst({
+        where: {
+          id: spaceMeta.mapId,
         },
+        select: {
+          mapElements: true,
+          width: true,
+          height: true,
+        },
+      });
+      console.log("after");
+      if (!map) {
+        throw new Error("map not found");
+      }
+      let space = await prisma.$transaction(async () => {
+        const space = await prisma.space.create({
+          data: {
+            name: spaceMeta.name,
+
+            width: parseInt(spaceMeta.dimensions.split("x")[0] || ""),
+            height: parseInt(spaceMeta.dimensions.split("x")[1] || ""),
+            creatorId: createrId,
+          },
+        });
+
+        await prisma.spaceElements.createMany({
+          data: map.mapElements.map((e) => ({
+            spaceId: space.id,
+            elementId: e.elementId,
+            x: e.x!,
+            y: e.y!,
+          })),
+        });
+
+        return space.id;
       });
       return space;
     } catch (error) {
       throw new Error("error while creating space");
     }
   }
-  async deleteSpace(spaceId: string) {
+  async deleteSpace(spaceId: string, userId: string) {
     try {
+      console.log("delete service is with" + spaceId + userId);
+      const space = await prisma.space.findUnique({
+        where: {
+          id: spaceId,
+        },
+        select: {
+          creatorId: true,
+        },
+      });
+      console.log(space);
+      if (!space) {
+        throw new Error("space does not  exist");
+      }
+
+      if (space.creatorId !== userId) {
+        console.log("code should reach here");
+
+        return "unautherized";
+      }
       const result = await prisma.space.delete({
         where: {
           id: spaceId,
         },
       });
+
       return { msg: "deleted" };
     } catch (error) {
+      console.error(error);
       throw new Error("error while deleting space");
     }
   }
@@ -39,16 +102,25 @@ export class spaceServices {
   }
   async getSpaceById(spaceId: string) {
     try {
-      const space = await prisma.space.findFirst({
+      const metaData = await prisma.space.findFirst({
         where: {
           id: spaceId,
         },
         select: {
           width: true,
           height: true,
-          SpaceElement: true,
+          elements: true,
         },
       });
+      if (!metaData) {
+        throw new Error("space does exits");
+        return;
+      }
+
+      const space = {
+        dimensions: metaData?.width + "x" + metaData?.height,
+        elements: metaData?.elements,
+      };
 
       return space;
     } catch (error) {
@@ -58,12 +130,35 @@ export class spaceServices {
   }
   async createSpaceElement(spaceMeta: speceElementMeta) {
     try {
-      const spaceElement = await prisma.spaceElement.create({
+      const space = (await prisma.space.findFirst({
+        where: {
+          id: spaceMeta.spaceId,
+        },
+      })) as {
+        id: string;
+        name: string;
+        width: number;
+        height: number;
+        thumbnail: string | null;
+        creatorId: string;
+      };
+      if (
+        space.width < spaceMeta.x ||
+        space.height < spaceMeta.y ||
+        spaceMeta.x < 0 ||
+        spaceMeta.y < 0
+      ) {
+        throw new Error("element's dimension is out of bound");
+      }
+      if (!space) {
+        throw new Error("space does exist");
+      }
+      const spaceElement = await prisma.spaceElements.create({
         data: {
           elementId: spaceMeta.elementId,
           spaceId: spaceMeta.spaceId,
-          x: spaceMeta.width,
-          y: spaceMeta.height,
+          x: spaceMeta.x,
+          y: spaceMeta.y,
         },
       });
       return spaceElement;
@@ -72,11 +167,11 @@ export class spaceServices {
       throw new Error("error while creating space element");
     }
   }
-  async deleteSpaceElement(elementMeta: { elementId: string }) {
+  async deleteSpaceElement(elementMeta: { id: string }) {
     try {
-      await prisma.spaceElement.delete({
+      await prisma.spaceElements.delete({
         where: {
-          id: elementMeta.elementId,
+          id: elementMeta.id,
         },
       });
